feat: new system configuration structure
