+++ module/config/chars.mjs
@@ -1,0 +1,5 @@
+export const CHARS = {
+	body: "METANTHROPES.CHARS.BODY.LABEL",
+	mind: "METANTHROPES.CHARS.MIND.LABEL",
+	soul: "METANTHROPES.CHARS.SOUL.LABEL",
+};
