+++ module/config/stats.mjs
@@ -1,0 +1,11 @@
+export const STATS = {
+	endurance: 'METANTHROPES.STATS.ENDURANCE.LABEL',
+	power: 'METANTHROPES.STATS.POWER.LABEL',
+	reflexes: 'METANTHROPES.STATS.REFLEXES.LABEL',
+	creativity: 'METANTHROPES.STATS.CREATIVITY.LABEL',
+	manipulation: 'METANTHROPES.STATS.MANIPULATION.LABEL',
+	perception: 'METANTHROPES.STATS.PERCEPTION.LABEL',
+	awareness: 'METANTHROPES.STATS.AWARENESS.LABEL',
+	willpower: 'METANTHROPES.STATS.WILLPOWER.LABEL',
+	consciousness: 'METANTHROPES.STATS.CONSIOUSNESS.LABEL',
+};
