+++ module/config/system.mjs
@@ -1,0 +1,10 @@
+import { CHARS } from "./chars.mjs";
+import { STATS } from "./stats.mjs";
+import { MOVEMENT } from "./movement.mjs";
+
+export const SYSTEM = {
+	id: "metanthropes",
+	CHARS,
+	STATS,
+	MOVEMENT,
+};
